Update allows dist/index.html to be served with very basic react components
live-sever should be installed for this manual test.
live-server uses default 127.0.0.1:8080 to serve the file
note also that in the App.jsx I created the Restaurantmeallist in the same file thus
avoiding modularization for now.  Note also, that I create it after the App component that uses it
I think this is ok.  Are there times where creating later could hurt the operations.

I did notice that the order of the components in the recastly were in opposite order of when the component would be used.
I.E. The app component which gets used first is in the last position.